requirements.txt
google_api_python_client==2.166.0
pandas==2.2.3
protobuf==6.31.0
streamlit==1.42.2
streamlit==1.39.0
